Cleanup refactors
