+++ src/lib.rs
@@ -208,6 +208,7 @@ use prelude::*;
 
 pub(crate) mod portable_par;
 
+pub mod bevy_compat;
 pub mod bundles;
 pub mod commands;
 pub mod floating_origins;
@@ -247,9 +248,9 @@ pub mod prelude {
     pub use world_query::{GridTransform, GridTransformOwned, GridTransformReadOnly};
 
     #[cfg(feature = "camera")]
-    pub use camera::{CameraController, CameraControllerPlugin};
+    pub use camera::{BigSpaceCameraControllerPlugin, CameraController};
     #[cfg(feature = "debug")]
-    pub use debug::FloatingOriginDebugPlugin;
+    pub use debug::BigSpaceDebugPlugin;
 }
 
 /// Contains the [`GridPrecision`] integer index type, which defines how much precision is available
